Approximate Bayesian posterior for hierarchical Normal model

Approximate a Bayesian posterior from a set ofCyclops likelihood profiles under a hierarchical normal model using the Markov chain Monte Carlo engine BEAST.

approximateHierarchicalNormalPosterior(
  likelihoodProfiles,
  chainLength = 1100000,
  burnIn = 1e+05,
  subSampleFrequency = 100,
  effectPriorMean = 0,
  effectPriorSd = 0.5,
  nu0 = 1,
  sigma0 = 1,
  effectStartingValue = 0,
  precisionStartingValue = 1,
  seed = 1
)

Arguments

likelihoodProfiles

List of grid likelihoods profiled with Cyclops.

chainLength

Number of MCMC iterations.

burnIn

Number of MCMC iterations to consider as burn in.

subSampleFrequency

Subsample frequency for the MCMC.

effectPriorMean

Prior mean for global parameter

effectPriorSd

Prior standard deviation for the global parameter

nu0

Prior "sample size" for precision (with precision ~ gamma(nu0/2, nu0*sigma0/2))

sigma0

Prior "variance" for precision (with precision ~ gamma(nu0/2, nu0*sigma0/2))

effectStartingValue

Initial value for global & local parameter

precisionStartingValue

Initial value for the precision

seed

Seed for the random number generator.

Value

A data frame with the point estimates and 95% credible intervals for the the global and local parameter, as well as the global precision. Attributes of the data frame contain the MCMC trace for diagnostics.

Examples

#TBD