Compute a Bayesian meta-analysis

Compute a Bayesian meta-analysis using the Markov chain Monte Carlo (MCMC) engine BEAST.

A normal and half-normal prior are used for the mu and tau parameters, respectively, with standard deviations as defined by the priorSd argument.

computeBayesianMetaAnalysis(
  data,
  chainLength = 1100000,
  burnIn = 1e+05,
  subSampleFrequency = 100,
  priorSd = c(2, 0.5),
  alpha = 0.05
)

Arguments

data
A data frame containing either normal, skew-normal, custom parametric, or grid likelihood data, with one row per database.

chainLength
Number of MCMC iterations.

burnIn
Number of MCMC iterations to consider as burn in.

subSampleFrequency
Subsample frequency for the MCMC.

priorSd
A two-dimensional vector with the standard deviation of the prior for mu and tau, respectively.

alpha
The alpha (expected type I error) used for the credible intervals.

Value

A data frame with the point estimates and 95% credible intervals for the mu and tau parameters (the mean and standard deviation of the distribution from which the per-site effect sizes are drawn). Attributes of the data frame contain the MCMC trace and the detected approximation type.

Examples

populations <- simulatePopulations()

fitModelInDatabase <- function(population) {
  cyclopsData <- Cyclops::createCyclopsData(Surv(time, y) ~ x + strata(stratumId),
                                            data = population,
                                            modelType = "cox")
  cyclopsFit <- Cyclops::fitCyclopsModel(cyclopsData)
  approximation <-  approximateLikelihood(cyclopsFit, "x")
  return(approximation)
}
approximations <- lapply(populations, fitModelInDatabase)
approximations <- do.call("rbind", approximations)
estimate <- computeBayesianMetaAnalysis(approximations)

#> Detected data following custom parameric distribution
#> Performing MCMC. This may take a while

estimate

#>          mu     mu95Lb   mu95Ub      muSe       tau     tau95Lb   tau95Ub
#> 1 0.4182657 -0.3517788 1.149729 0.3772438 0.2852941 0.001675841 0.8201178

# mu     mu95Lb    mu95Ub     muSe        tau      tau95Lb  tau95Ub
# 0.0003129562 -0.1747429 0.1723472 0.089661 0.07759992 0.0002024991 0.301007