Generate settings for the Bayesian random-effects hierarchical meta-analysis model

This function generates a settings list for fitting a Bayesian hierarchical meta-analysis model. See computeHierarchicalMetaAnalysis() for more details.

generateBayesianHMAsettings(
  primaryEffectPriorStd = 1,
  secondaryEffectPriorStd = 1,
  globalExposureEffectPriorMean = c(0),
  globalExposureEffectPriorStd = c(2),
  primaryEffectPrecisionPrior = c(1, 1),
  secondaryEffectPrecisionPrior = c(1, 1),
  errorPrecisionPrior = c(1, 1),
  errorPrecisionStartValue = 1,
  includeSourceEffect = TRUE,
  includeExposureEffect = TRUE,
  exposureEffectCount = 1,
  separateExposurePrior = FALSE,
  chainLength = 1100000,
  burnIn = 1e+05,
  subSampleFrequency = 100
)

Arguments

primaryEffectPriorStd

Standard deviation for the average outcome effect.

secondaryEffectPriorStd

Standard deviation for the average data-source effect.

globalExposureEffectPriorMean

Prior mean for the global main exposure effect; can be a multiple entry vector if there are multiple outcomes of interest

globalExposureEffectPriorStd

Prior standard deviation for the global main exposure effect; can be a multiple entry vector if there are multiple outcomes of interest

primaryEffectPrecisionPrior

Shape and scale for the gamma prior of the precision term in the random effects model (normal) for individual outcome effects.

secondaryEffectPrecisionPrior

Shape and scale for the gamma prior of the precision term in the random effects model (normal) for individual data-source effects.

errorPrecisionPrior

Shape and scale for the gamma prior of the precision term in the normal model for random errors.

errorPrecisionStartValue

Initial value for the error distribution's precision term.

includeSourceEffect

Whether or not to consider the data-source-specific (secondary) random effects. Default is TRUE.

includeExposureEffect

Whether or not to estimate the main effect of interest. Default is TRUE.

exposureEffectCount

Number of main outcomes of interest to estimate effect for? Default = 1

separateExposurePrior

Use a separable prior on the main exposure effect? Default is FALSE.

chainLength

Number of MCMC iterations.

burnIn

Number of MCMC iterations to consider as burn in.

subSampleFrequency

Subsample ("thinning") frequency for the MCMC.

Value

A list with all the settings to use in the computeHierarchicalMetaAnalysis() function.

See also